fix: keep CSS merging dependencies out of navigation bundle
